refactor: Use packaging by feature
