+++ shuttle-mobile/app/src/main/java/com/example/shuttlemobile/driver/DriverMainActivity.java
@@ -1,4 +1,4 @@
-package com.example.shuttlemobile.activity.passenger;
+package com.example.shuttlemobile.driver;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.example.shuttlemobile.R;
 
-public class PassengerMainActivity extends AppCompatActivity {
+public class DriverMainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_passenger_main);
+        setContentView(R.layout.activity_driver_main);
     }
 }
