+++ shuttle-mobile/app/src/main/res/layout/activity_driver_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.driver.DriverMainActivity">
+    tools:context=".driver.DriverMainActivity">
 
     <ToggleButton
         android:id="@+id/driver_main_activity_toggle"
