+++ shuttle-mobile/app/src/main/res/layout/activity_passenger_main.xml
@@ -4,4 +4,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.passenger.PassengerMainActivity"/>
+    tools:context=".passenger.PassengerMainActivity"/>
