+++ shuttle-mobile/app/src/main/res/layout/activity_user_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.UserLoginActivity">
+    tools:context=".user.UserLoginActivity">
 
     <TextView
         android:id="@+id/login_label"
